[REF] extract label properties and functions to mixin
